auth/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
google-api-python-client==1.10.0
google-auth==1.20.1
git+https://github.com/hfaran/piazza-api.git@develop
-r common/requirements.txt
